+++ pypsa/optimization/constraints.py
@@ -8,7 +8,7 @@ import logging
 import linopy
 import pandas as pd
 from linopy import LinearExpression, merge
-from numpy import inf, isfinite
+from numpy import inf, isfinite, round
 from scipy import sparse
 from xarray import DataArray, Dataset, concat
 
@@ -25,12 +25,12 @@ from pypsa.optimization.common import reindex
 logger = logging.getLogger(__name__)
 
 
-def define_operational_constraints_for_non_extendables(
+def define_operational_constraints_for_non_extendables_and_non_committables(
     n, sns, c, attr, transmission_losses
 ):
     """
     Sets power dispatch constraints for non-extendable and non-commitable
-    assets for a given component and a given attribute.
+    assets for a given component and a given attribute,  whether they are modular or not.
 
     Parameters
     ----------
@@ -62,19 +62,19 @@ def define_operational_constraints_for_non_extendables(
         dispatch_lower = (1, dispatch_lower), (-1, loss)
         dispatch_upper = (1, dispatch_upper), (1, loss)
     n.model.add_constraints(
-        dispatch_lower, ">=", lower, name=f"{c}-fix-{attr}-lower", mask=active
+        dispatch_lower, ">=", lower, name=f"{c}-fix-non-comm-{attr}-lower", mask=active
     )
     n.model.add_constraints(
-        dispatch_upper, "<=", upper, name=f"{c}-fix-{attr}-upper", mask=active
+        dispatch_upper, "<=", upper, name=f"{c}-fix-non-comm-{attr}-upper", mask=active
     )
 
 
-def define_operational_constraints_for_extendables(
+def define_operational_constraints_for_extendables_but_non_committables(
     n, sns, c, attr, transmission_losses
 ):
     """
-    Sets power dispatch constraints for extendable devices for a given
-    component and a given attribute.
+    Sets power dispatch constraints for extendable but non commitable devices for a given
+    component and a given attribute,  whether they are modular or not.
 
     Parameters
     ----------
@@ -87,7 +87,8 @@ def define_operational_constraints_for_extendables(
         name of the attribute, e.g. 'p'
     """
     ext_i = n.get_extendable_i(c)
-
+    ext_i = ext_i.difference(n.get_committable_i(c)).rename(ext_i.name)
+    
     if ext_i.empty:
         return
 
@@ -105,16 +106,409 @@ def define_operational_constraints_for_extendables(
         lhs_upper += ((1, loss),)
 
     n.model.add_constraints(
-        lhs_lower, ">=", 0, name=f"{c}-ext-{attr}-lower", mask=active
+        lhs_lower, ">=", 0, name=f"{c}-ext-non-comm-{attr}-lower", mask=active
     )
     n.model.add_constraints(
-        lhs_upper, "<=", 0, name=f"{c}-ext-{attr}-upper", mask=active
+        lhs_upper, "<=", 0, name=f"{c}-ext-non-comm-{attr}-upper", mask=active
     )
 
-
-def define_operational_constraints_for_committables(n, sns, c):
+def  define_operational_constraints_for_extendables_and_committables_but_non_modular(
+    n, sns, c, attr, transmission_losses
+):
     """
-    Sets power dispatch constraints for committable devices for a given
+    Sets power dispatch constraints for extendable, committable, but non modular devices for a given
+    component and a given attribute. Extendability is preferred over committability (committability
+    is simply ignored).
+
+    Parameters
+    ----------
+    n : pypsa.Network
+    sns : pd.Index
+        Snapshots of the constraint.
+    c : str
+        name of the network component
+    attr : str
+        name of the attribute, e.g. 'p'
+    """
+    ext_i = n.get_extendable_i(c)
+    com_i = n.get_committable_i(c)
+    not_mod_i = n.df(c).query(f"({nominal_attrs[c]}_mod==0)").index
+
+    ext_i = ext_i.intersection(com_i).intersection(not_mod_i).rename(ext_i.name)
+    
+    if ext_i.empty:
+        return
+
+    min_pu, max_pu = map(DataArray, get_bounds_pu(n, c, sns, ext_i, attr))
+    dispatch = reindex(n.model[f"{c}-{attr}"], c, ext_i)
+    capacity = n.model[f"{c}-{nominal_attrs[c]}"]
+
+    active = get_activity_mask(n, c, sns, ext_i) if n._multi_invest else None
+
+    lhs_lower = (1, dispatch), (-min_pu, capacity)
+    lhs_upper = (1, dispatch), (-max_pu, capacity)
+    if c in n.passive_branch_components and transmission_losses:
+        loss = reindex(n.model[f"{c}-loss"], c, ext_i)
+        lhs_lower += ((-1, loss),)
+        lhs_upper += ((1, loss),)
+
+    n.model.add_constraints(
+        lhs_lower, ">=", 0, name=f"{c}-ext-comm-non-mod-{attr}-lower", mask=active
+    )
+    n.model.add_constraints(
+        lhs_upper, "<=", 0, name=f"{c}-ext-comm-non-mod-{attr}-upper", mask=active
+    )
+
+def define_committability_variables_constraints_for_non_modular(n, sns, c, attr):
+    """
+    This function sets the upper limit of committable variables (status, start-up, shut-down) to 1 for all
+    committable components for which modularity is not used, regardless of whether they are extendable or not.
+    _____
+
+    Parameters
+    ----------
+    n : pypsa.Network
+    sns : pd.Index
+        Snapshots of the constraint.
+    c : str
+        name of the network component
+    attr : str
+        name of the attribute, e.g. 'p_nom'
+    """
+    com_i = n.get_committable_i(c)
+    not_mod_i = n.df(c).query(f"({attr}_mod==0)").index
+    com_i = com_i.intersection(not_mod_i).rename(com_i.name)
+
+    if com_i.empty:
+        return
+
+    m = n.model
+    ext_i = n.get_extendable_i(c)
+    active = get_activity_mask(n, c, sns, ext_i) if n._multi_invest else None
+
+    status = m.variables[f"{c}-status"].loc[sns,com_i]
+    n.model.add_constraints(
+        status, "<=", 1, name=f"{c}-status-{attr}-upper-nonmod", mask=active
+    )
+    start_up = m.variables[f"{c}-start_up"].loc[sns, com_i]
+    n.model.add_constraints(
+        start_up, "<=", 1, name=f"{c}-start_up-{attr}-upper-nonmod", mask=active
+    )
+
+    shut_down = m.variables[f"{c}-shut_down"].loc[sns,com_i]
+    n.model.add_constraints(
+        shut_down, "<=", 1, name=f"{c}-shut_down-{attr}-upper-nonmod", mask=active
+    )
+
+def define_committability_variables_constraints_for_modular_and_non_extendables(n, sns, c, attr):
+    """
+    This function sets the upper limit of committable variables (status, start-up, shut-down) to the
+    installed number of modules for all committable and non-extendable components. The number of modules
+    is calculated as the nominal power divided by the nominal dimension of the individual module (e.g. p_nom/p_nom_mod).
+
+    If the number of modules is not an integer number, the function returns an error message.
+    _____
+
+    Parameters
+    ----------
+    n : pypsa.Network
+    sns : pd.Index
+        Snapshots of the constraint.
+    c : str
+        name of the network component
+    attr : str
+        name of the attribute, e.g. 'p_nom'
+    """
+    com_i = n.get_committable_i(c)
+    fix_i = n.df(c).query(f"not {attr}_extendable").index
+    mod_i = n.df(c).query(f"({attr}_mod>0)").index
+
+    inter_i = com_i.intersection(mod_i).intersection(fix_i).rename(com_i.name)
+    
+    if inter_i.empty:
+        return
+
+    m = n.model
+    
+    ext_i = n.get_extendable_i(c)
+    active = get_activity_mask(n, c, sns, ext_i) if n._multi_invest else None
+
+    n_mod = n.df(c)[attr].loc[inter_i]/n.df(c)[attr + "_mod"].loc[inter_i]
+    diff_n_mod = abs(n_mod - round(n_mod))
+    non_integers_n_mod_i = diff_n_mod[diff_n_mod > 10**-6].index
+    
+    if  not non_integers_n_mod_i.empty:
+        msg = (
+        "For non-extendable but committable assets, if both p_nom and p_nom_mod are declared, p_nom must be a"
+        "must be a multiple of p_nom_mod. Found "
+                f"assets in component {c} do not respect this criterion:"
+                f"\n\n\t{', '.join(non_integers_n_mod_i)}"
+        )
+        raise ValueError(msg)
+    
+    
+    rhs = pd.DataFrame(0, sns, inter_i)
+    rhs.loc[sns, inter_i] = n_mod.loc[inter_i].values
+
+    status = m.variables[f"{c}-status"].loc[sns,inter_i]
+    n.model.add_constraints(
+        status, "<=", rhs, name=f"{c}-status-{attr}-upper-mod_nonext", mask=active
+    )
+
+    start_up = m.variables[f"{c}-start_up"].loc[sns, inter_i]
+    n.model.add_constraints(
+        start_up, "<=", rhs, name=f"{c}-start_up-{attr}-upper-mod_nonext", mask=active
+    )
+
+    shut_down = m.variables[f"{c}-shut_down"].loc[sns, inter_i]
+    n.model.add_constraints(
+        shut_down, "<=", rhs, name=f"{c}-shut_down-{attr}-upper-mod_nonext", mask=active
+    )
+
+def define_committability_variables_constraints_for_modular_and_extendables(n, sns, c, attr):
+    """
+    This function sets the upper limit of committable variables (status, start-up, shut-down) to the
+    variable n_mod for all committable, extendable and modular components. 
+    _____
+
+    Parameters
+    ----------
+    n : pypsa.Network
+    sns : pd.Index
+        Snapshots of the constraint.
+    c : str
+        name of the network component
+    attr : str
+        name of the attribute, e.g. 'p_nom'
+    """
+    com_i = n.get_committable_i(c)
+    ext_i = n.df(c).query(f"{attr}_extendable").index
+    mod_i = n.df(c).query(f"({attr}_mod>0)").index
+
+    inter_i = com_i.intersection(mod_i).intersection(ext_i).rename(com_i.name)
+
+    if inter_i .empty:
+        return
+
+    m = n.model
+    
+    
+    active = get_activity_mask(n, c, sns, ext_i) if n._multi_invest else None
+
+    n_mod = n.model[f"{c}-n_mod"].loc[inter_i]
+    
+    #ALLERT DIMENSIONI DISEQUAZIONE
+    status = m.variables[f"{c}-status"].loc[sns,inter_i]
+    lhs = ((1, status), (-1, n_mod))
+    n.model.add_constraints(
+        lhs, "<=", 0, name=f"{c}-status-{attr}-upper-mod_ext", mask=active
+    )
+
+    #ALLERT DIMENSIONI DISEQUAZIONE
+    start_up = m.variables[f"{c}-start_up"].loc[sns,inter_i]
+    lhs = ((1, start_up), (-1, n_mod))
+    n.model.add_constraints(
+        lhs, "<=", 0, name=f"{c}-start_up-{attr}-upper-mod_ext", mask=active
+    )
+    
+    #ALLERT DIMENSIONI DISEQUAZIONE
+    shut_down = m.variables[f"{c}-shut_down"].loc[sns,inter_i]
+    lhs = ((1, shut_down), (-1, n_mod))
+    n.model.add_constraints(
+        lhs, "<=", 0, name=f"{c}-shut_down-{attr}-upper-mod_ext", mask=active
+    )
+    
+def define_operational_constraints_for_committables_and_modular(n, sns, c):
+    """
+    Sets power dispatch constraints for committable and modular devices for a given
+    component and a given attribute, whether they are extendable or not. The linearized
+    approximation of the unit commitment problem is inspired by Hua et al. (2017) DOI:
+    10.1109/TPWRS.2017.2735026.
+
+    Parameters
+    ----------
+    n : pypsa.Network
+    sns : pd.Index
+        Snapshots of the constraint.
+    c : str
+        name of the network component
+    """
+    name_func = "-committable-modular"
+
+    com_i = n.get_committable_i(c)
+    mod_i = n.df(c).query(f"({nominal_attrs[c]}_mod>0)").index
+    com_i = com_i.intersection(mod_i).rename(com_i.name)
+
+    if com_i.empty:
+        return
+    
+    # variables
+    status = n.model[f"{c}-status"].loc[sns, com_i]
+    start_up = n.model[f"{c}-start_up"].loc[sns, com_i]
+    shut_down = n.model[f"{c}-shut_down"].loc[sns, com_i]
+    status_diff = status - status.shift(snapshot=1)
+    p = reindex(n.model[f"{c}-p"], c, com_i)
+    
+    active = get_activity_mask(n, c, sns, com_i) if n._multi_invest else None
+
+    # parameters
+    nominal = DataArray(n.df(c)[nominal_attrs[c]+'_mod'].reindex(com_i))
+    min_pu, max_pu = map(DataArray, get_bounds_pu(n, c, sns, com_i, "p"))
+    lower_p = min_pu * nominal
+    upper_p = max_pu * nominal
+    min_up_time_set = n.df(c).min_up_time[com_i]
+    min_down_time_set = n.df(c).min_down_time[com_i]
+    ramp_up_limit = nominal * n.df(c).ramp_limit_up[com_i].fillna(1)
+    ramp_down_limit = nominal * n.df(c).ramp_limit_down[com_i].fillna(1)
+    ramp_start_up = nominal * n.df(c).ramp_limit_start_up[com_i]
+    ramp_shut_down = nominal * n.df(c).ramp_limit_shut_down[com_i]
+    up_time_before_set = n.df(c)["up_time_before"].reindex(com_i)
+    down_time_before_set = n.df(c)["down_time_before"].reindex(com_i)
+    initially_up = up_time_before_set.astype(bool)
+    initially_down = down_time_before_set.astype(bool)
+
+    # check if there are status calculated/fixed before given sns interval
+    if sns[0] != n.snapshots[0]:
+        start_i = n.snapshots.get_loc(sns[0])
+        # get generators which are online until the first regarded snapshot
+        until_start_up = n.pnl(c).status.iloc[:start_i][::-1].reindex(columns=com_i)
+        ref = range(1, len(until_start_up) + 1)
+        up_time_before = until_start_up[until_start_up.cumsum().eq(ref, axis=0)].sum()
+        up_time_before_set = up_time_before.clip(upper=min_up_time_set)
+        initially_up = up_time_before_set.astype(bool)
+        # get number of snapshots for generators which are offline before the first regarded snapshot
+        until_start_down = ~until_start_up.astype(bool)
+        ref = range(1, len(until_start_down) + 1)
+        down_time_before = until_start_down[
+            until_start_down.cumsum().eq(ref, axis=0)
+        ].sum()
+        down_time_before_set = down_time_before.clip(upper=min_down_time_set)
+        initially_down = down_time_before_set.astype(bool)
+
+    # lower dispatch level limit
+    lhs = (1, p), (-lower_p, status)
+    n.model.add_constraints(lhs, ">=", 0, name=f"{c}-com-mod-p-lower"+name_func, mask=active)
+
+    # upper dispatch level limit
+    lhs = (1, p), (-upper_p, status)
+    n.model.add_constraints(lhs, "<=", 0, name=f"{c}-com-mod-p-upper"+name_func, mask=active)
+
+    # state-transition constraint
+    rhs = pd.DataFrame(0, sns, com_i)
+    rhs.loc[sns[0], initially_up] = -1
+    lhs = start_up - status_diff
+    n.model.add_constraints(
+        lhs, ">=", rhs, name=f"{c}-com-mod-transition-start-up"+name_func, mask=active
+    )
+
+    rhs = pd.DataFrame(0, sns, com_i)
+    rhs.loc[sns[0], initially_up] = 1
+    lhs = shut_down + status_diff
+    n.model.add_constraints(
+        lhs, ">=", rhs, name=f"{c}-com-mod-transition-shut-down"+name_func, mask=active
+    )
+
+    # min up time
+    mask = get_activity_mask(n, c, sns[1:], com_i)
+    expr = []
+    min_up_time_i = com_i[min_up_time_set.astype(bool)]
+    if not min_up_time_i.empty:
+        for g in min_up_time_i:
+            su = start_up.loc[:, g]
+            expr.append(su.rolling(snapshot=min_up_time_set[g]).sum())
+        lhs = -status.loc[:, min_up_time_i] + merge(expr, dim=com_i.name)
+        lhs = lhs.sel(snapshot=sns[1:])
+        n.model.add_constraints(
+            lhs, "<=", 0, name=f"{c}-com-mod-up-time"+name_func, mask=mask[min_up_time_i]
+        )
+
+    # min down time
+    expr = []
+    min_down_time_i = com_i[min_down_time_set.astype(bool)]
+    if not min_down_time_i.empty:
+        for g in min_down_time_i:
+            su = shut_down.loc[:, g]
+            expr.append(su.rolling(snapshot=min_down_time_set[g]).sum())
+        lhs = status.loc[:, min_down_time_i] + merge(expr, dim=com_i.name)
+        lhs = lhs.sel(snapshot=sns[1:])
+        n.model.add_constraints(
+            lhs, "<=", 1, name=f"{c}-com-mod-down-time"+name_func, mask=mask[min_down_time_i]
+        )
+
+    # up time before
+    timesteps = pd.DataFrame([range(1, len(sns) + 1)] * len(com_i), com_i, sns).T
+    if initially_up.any():
+        must_stay_up = (min_up_time_set - up_time_before_set).clip(lower=0)
+        mask = (must_stay_up >= timesteps) & initially_up
+        name = f"{c}-com-mod-status-min_up_time_must_stay_up"+name_func
+        mask = mask & active if active is not None else mask
+        n.model.add_constraints(status, "=", 1, name=name, mask=mask)
+
+    # down time before
+    if initially_down.any():
+        must_stay_down = (min_down_time_set - down_time_before_set).clip(lower=0)
+        mask = (must_stay_down >= timesteps) & initially_down
+        name = f"{c}-com-mod-status-min_down_time_must_stay_up"+name_func
+        mask = mask & active if active is not None else mask
+        n.model.add_constraints(status, "=", 0, name=name, mask=mask)
+
+    # linearized approximation because committable can partly start up and shut down
+    cost_equal = all(
+        n.df(c).loc[com_i, "start_up_cost"] == n.df(c).loc[com_i, "shut_down_cost"]
+    )
+    # only valid additional constraints if start up costs equal to shut down costs
+    if n._linearized_uc and not cost_equal:
+        logger.warning(
+            "The linear relaxation of the unit commitment cannot be "
+            "tightened since the start up costs are not equal to the "
+            "shut down costs. Proceed with the linear relaxation "
+            "without the tightening by additional constraints. "
+            "This might result in a longer solving time."
+        )
+    if n._linearized_uc and cost_equal:
+        # dispatch limit for partly start up/shut down for t-1
+        lhs = (
+            p.shift(snapshot=1)
+            - ramp_shut_down * status.shift(snapshot=1)
+            - (upper_p - ramp_shut_down) * (status - start_up)
+        )
+        lhs = lhs.sel(snapshot=sns[1:])
+        n.model.add_constraints(lhs, "<=", 0, name=f"{c}-com-mod-p-before"+name_func, mask=active)
+
+        # dispatch limit for partly start up/shut down for t
+        lhs = p - upper_p * status + (upper_p - ramp_start_up) * start_up
+        lhs = lhs.sel(snapshot=sns[1:])
+        n.model.add_constraints(lhs, "<=", 0, name=f"{c}-com-mod-p-current"+name_func, mask=active)
+
+        # ramp up if committable is only partly active and some capacity is starting up
+        lhs = (
+            p
+            - p.shift(snapshot=1)
+            - (lower_p + ramp_up_limit) * status
+            + lower_p * status.shift(snapshot=1)
+            + (lower_p + ramp_up_limit - ramp_start_up) * start_up
+        )
+        lhs = lhs.sel(snapshot=sns[1:])
+        n.model.add_constraints(
+            lhs, "<=", 0, name=f"{c}-com-mod-partly-start-up"+name_func, mask=active
+        )
+
+        # ramp down if committable is only partly active and some capacity is shutting up
+        lhs = (
+            p.shift(snapshot=1)
+            - p
+            - ramp_shut_down * status.shift(snapshot=1)
+            + (ramp_shut_down - ramp_down_limit) * status
+            - (lower_p + ramp_down_limit - ramp_shut_down) * start_up
+        )
+        lhs = lhs.sel(snapshot=sns[1:])
+        n.model.add_constraints(
+            lhs, "<=", 0, name=f"{c}-com-mod-partly-shut-down"+name_func, mask=active
+        )
+
+def define_operational_constraints_for_committables_non_modular_non_extendables(n, sns, c):
+    """
+    Sets power dispatch constraints for committable but non modular and non extendable devices for a given
     component and a given attribute. The linearized approximation of the unit
     commitment problem is inspired by Hua et al. (2017) DOI:
     10.1109/TPWRS.2017.2735026.
@@ -128,14 +522,18 @@ def define_operational_constraints_for_committables(n, sns, c):
         name of the network component
     """
     com_i = n.get_committable_i(c)
+    not_mod_i = n.df(c).query(f"({nominal_attrs[c]}_mod==0)").index
+    fix_i = n.get_non_extendable_i(c)
+
+    com_i = com_i.intersection(not_mod_i).intersection(fix_i).rename(com_i.name)
 
     if com_i.empty:
         return
 
     # variables
-    status = n.model[f"{c}-status"]
-    start_up = n.model[f"{c}-start_up"]
-    shut_down = n.model[f"{c}-shut_down"]
+    status = n.model[f"{c}-status"].loc[sns, com_i]
+    start_up = n.model[f"{c}-start_up"].loc[sns, com_i]
+    shut_down = n.model[f"{c}-shut_down"].loc[sns, com_i]
     status_diff = status - status.shift(snapshot=1)
     p = reindex(n.model[f"{c}-p"], c, com_i)
     active = get_activity_mask(n, c, sns, com_i) if n._multi_invest else None
@@ -176,25 +574,25 @@ def define_operational_constraints_for_committables(n, sns, c):
 
     # lower dispatch level limit
     lhs = (1, p), (-lower_p, status)
-    n.model.add_constraints(lhs, ">=", 0, name=f"{c}-com-p-lower", mask=active)
+    n.model.add_constraints(lhs, ">=", 0, name=f"{c}-com-non-mod-fix-p-lower", mask=active)
 
     # upper dispatch level limit
     lhs = (1, p), (-upper_p, status)
-    n.model.add_constraints(lhs, "<=", 0, name=f"{c}-com-p-upper", mask=active)
+    n.model.add_constraints(lhs, "<=", 0, name=f"{c}-com-non-mod-fix-p-upper", mask=active)
 
     # state-transition constraint
     rhs = pd.DataFrame(0, sns, com_i)
     rhs.loc[sns[0], initially_up] = -1
     lhs = start_up - status_diff
     n.model.add_constraints(
-        lhs, ">=", rhs, name=f"{c}-com-transition-start-up", mask=active
+        lhs, ">=", rhs, name=f"{c}-com-non-mod-fix-transition-start-up", mask=active
     )
 
     rhs = pd.DataFrame(0, sns, com_i)
     rhs.loc[sns[0], initially_up] = 1
     lhs = shut_down + status_diff
     n.model.add_constraints(
-        lhs, ">=", rhs, name=f"{c}-com-transition-shut-down", mask=active
+        lhs, ">=", rhs, name=f"{c}-com-non-mod-fix-transition-shut-down", mask=active
     )
 
     # min up time
@@ -208,7 +606,7 @@ def define_operational_constraints_for_committables(n, sns, c):
         lhs = -status.loc[:, min_up_time_i] + merge(expr, dim=com_i.name)
         lhs = lhs.sel(snapshot=sns[1:])
         n.model.add_constraints(
-            lhs, "<=", 0, name=f"{c}-com-up-time", mask=mask[min_up_time_i]
+            lhs, "<=", 0, name=f"{c}-com-non-mod-fix-up-time", mask=mask[min_up_time_i]
         )
 
     # min down time
@@ -221,23 +619,25 @@ def define_operational_constraints_for_committables(n, sns, c):
         lhs = status.loc[:, min_down_time_i] + merge(expr, dim=com_i.name)
         lhs = lhs.sel(snapshot=sns[1:])
         n.model.add_constraints(
-            lhs, "<=", 1, name=f"{c}-com-down-time", mask=mask[min_down_time_i]
+            lhs, "<=", 1, name=f"{c}-com-non-mod-fix-down-time", mask=mask[min_down_time_i]
         )
 
+    #ATT: DA RIVEDERE  MODIFICARE
     # up time before
     timesteps = pd.DataFrame([range(1, len(sns) + 1)] * len(com_i), com_i, sns).T
     if initially_up.any():
         must_stay_up = (min_up_time_set - up_time_before_set).clip(lower=0)
         mask = (must_stay_up >= timesteps) & initially_up
-        name = f"{c}-com-status-min_up_time_must_stay_up"
+        name = f"{c}-com-non-mod-fix-status-min_up_time_must_stay_up"
         mask = mask & active if active is not None else mask
         n.model.add_constraints(status, "=", 1, name=name, mask=mask)
 
+    #ATT: DA RIVEDERE  MODIFICARE
     # down time before
     if initially_down.any():
         must_stay_down = (min_down_time_set - down_time_before_set).clip(lower=0)
         mask = (must_stay_down >= timesteps) & initially_down
-        name = f"{c}-com-status-min_down_time_must_stay_up"
+        name = f"{c}-com-non-mod-fix-status-min_down_time_must_stay_up"
         mask = mask & active if active is not None else mask
         n.model.add_constraints(status, "=", 0, name=name, mask=mask)
 
@@ -262,12 +662,12 @@ def define_operational_constraints_for_committables(n, sns, c):
             - (upper_p - ramp_shut_down) * (status - start_up)
         )
         lhs = lhs.sel(snapshot=sns[1:])
-        n.model.add_constraints(lhs, "<=", 0, name=f"{c}-com-p-before", mask=active)
+        n.model.add_constraints(lhs, "<=", 0, name=f"{c}-com-non-mod-fix-p-before", mask=active)
 
         # dispatch limit for partly start up/shut down for t
         lhs = p - upper_p * status + (upper_p - ramp_start_up) * start_up
         lhs = lhs.sel(snapshot=sns[1:])
-        n.model.add_constraints(lhs, "<=", 0, name=f"{c}-com-p-current", mask=active)
+        n.model.add_constraints(lhs, "<=", 0, name=f"{c}-com-non-mod-fix-p-current", mask=active)
 
         # ramp up if committable is only partly active and some capacity is starting up
         lhs = (
@@ -279,7 +679,7 @@ def define_operational_constraints_for_committables(n, sns, c):
         )
         lhs = lhs.sel(snapshot=sns[1:])
         n.model.add_constraints(
-            lhs, "<=", 0, name=f"{c}-com-partly-start-up", mask=active
+            lhs, "<=", 0, name=f"{c}-com-non-mod-fix-partly-start-up", mask=active
         )
 
         # ramp down if committable is only partly active and some capacity is shutting up
@@ -292,7 +692,7 @@ def define_operational_constraints_for_committables(n, sns, c):
         )
         lhs = lhs.sel(snapshot=sns[1:])
         n.model.add_constraints(
-            lhs, "<=", 0, name=f"{c}-com-partly-shut-down", mask=active
+            lhs, "<=", 0, name=f"{c}-com-non-mod-fix-partly-shut-down", mask=active
         )
 
 
@@ -386,8 +786,10 @@ def define_ramp_limit_constraints(n, sns, c, attr):
     fix_i = fix_i.difference(com_i).rename(fix_i.name)
     ext_i = n.get_extendable_i(c)
 
-    # ----------------------------- Fixed Generators ----------------------------- #
-
+    # ------------------- Fixed and non committable Generators (modular and non modular) ----------------------------- #
+    com_i = n.get_committable_i(c)
+    fix_i = n.get_non_extendable_i(c)
+    fix_i = fix_i.difference(com_i).rename(fix_i.name)
     assets = n.df(c).reindex(fix_i)
 
     p_nom = n.df(c)[nominal_attrs[c]].reindex(fix_i)
@@ -402,7 +804,7 @@ def define_ramp_limit_constraints(n, sns, c, attr):
             active.index, columns=fix_i
         )
         m.add_constraints(
-            lhs, "<=", rhs, name=f"{c}-fix-{attr}-ramp_limit_up", mask=mask
+            lhs, "<=", rhs, name=f"{c}-fix-non-comm-{attr}-ramp_limit_up", mask=mask
         )
 
     # fix down
@@ -415,16 +817,20 @@ def define_ramp_limit_constraints(n, sns, c, attr):
             active.index, columns=fix_i
         )
         m.add_constraints(
-            lhs, ">=", rhs, name=f"{c}-fix-{attr}-ramp_limit_down", mask=mask
+            lhs, ">=", rhs, name=f"{c}-fix-non-comm-{attr}-ramp_limit_down", mask=mask
         )
 
-    # ----------------------------- Extendable Generators ----------------------------- #
+    # ----------------------- Extendable but non committable Generators ----------------------------- #
 
+    ext_i = n.get_extendable_i(c)
+    com_i = n.get_committable_i(c)
+    ext_i = ext_i.difference(com_i).rename(ext_i.name)
+    
     assets = n.df(c).reindex(ext_i)
 
     # ext up
     if not ramp_limit_up[ext_i].isnull().all().all():
-        p_nom = m[f"{c}-p_nom"]
+        p_nom = m[f"{c}-p_nom"].loc[ext_i]
         limit_pu = DataArray(ramp_limit_up.reindex(active.index, columns=ext_i))
         lhs = p_actual(ext_i) - p_previous(ext_i) - limit_pu * p_nom
         rhs = rhs_start.reindex(columns=ext_i)
@@ -432,12 +838,12 @@ def define_ramp_limit_constraints(n, sns, c, attr):
             active.index, columns=ext_i
         )
         m.add_constraints(
-            lhs, "<=", rhs, name=f"{c}-ext-{attr}-ramp_limit_up", mask=mask
+            lhs, "<=", rhs, name=f"{c}-ext-non-comm-{attr}-ramp_limit_up", mask=mask
         )
 
     # ext down
     if not ramp_limit_down[ext_i].isnull().all().all():
-        p_nom = m[f"{c}-p_nom"]
+        p_nom = m[f"{c}-p_nom"].loc[ext_i]
         limit_pu = DataArray(ramp_limit_down.reindex(active.index, columns=ext_i))
         lhs = p_actual(ext_i) - p_previous(ext_i) + limit_pu * p_nom
         rhs = rhs_start.reindex(columns=ext_i)
@@ -445,11 +851,51 @@ def define_ramp_limit_constraints(n, sns, c, attr):
             active.index, columns=ext_i
         )
         m.add_constraints(
-            lhs, ">=", rhs, name=f"{c}-ext-{attr}-ramp_limit_down", mask=mask
+            lhs, ">=", rhs, name=f"{c}-ext-non-comm-{attr}-ramp_limit_down", mask=mask
         )
 
-    # ----------------------------- Committable Generators ----------------------------- #
+    # ----------------------- Extendable, committable but non modular Generators ----------------------------- #
 
+    ext_i = n.get_extendable_i(c)
+    com_i = n.get_committable_i(c)
+    not_mod_i = n.df(c).query(f"({nominal_attrs[c]}_mod==0)").index
+    ext_i = ext_i.intersection(com_i).intersection(not_mod_i).rename(ext_i.name)
+    
+    assets = n.df(c).reindex(ext_i)
+
+    # ext up
+    if not ramp_limit_up[ext_i].isnull().all().all():
+        p_nom = m[f"{c}-p_nom"].loc[ext_i]
+        limit_pu = DataArray(ramp_limit_up.reindex(active.index, columns=ext_i))
+        lhs = p_actual(ext_i) - p_previous(ext_i) - limit_pu * p_nom
+        rhs = rhs_start.reindex(columns=ext_i)
+        mask = active.reindex(columns=ext_i) & ~ramp_limit_up.isnull().reindex(
+            active.index, columns=ext_i
+        )
+        m.add_constraints(
+            lhs, "<=", rhs, name=f"{c}-ext-comm-non-mod-{attr}-ramp_limit_up", mask=mask
+        )
+
+    # ext down
+    if not ramp_limit_down[ext_i].isnull().all().all():
+        p_nom = m[f"{c}-p_nom"].loc[ext_i]
+        limit_pu = DataArray(ramp_limit_down.reindex(active.index, columns=ext_i))
+        lhs = p_actual(ext_i) - p_previous(ext_i) + limit_pu * p_nom
+        rhs = rhs_start.reindex(columns=ext_i)
+        mask = active.reindex(columns=ext_i) & ~ramp_limit_down.isnull().reindex(
+            active.index, columns=ext_i
+        )
+        m.add_constraints(
+            lhs, ">=", rhs, name=f"{c}-ext-comm-non-mod-{attr}-ramp_limit_down", mask=mask
+        )
+
+    # ---------------- Committable but non extendable and non modular Generators ----------------------------- #
+
+    com_i = n.get_committable_i(c)
+    fix_i = n.get_non_extendable_i(c)
+    not_mod_i = n.df(c).query(f"({nominal_attrs[c]}_mod==0)").index
+    com_i = com_i.intersection(fix_i).intersection(not_mod_i).rename(com_i.name)
+    
     assets = n.df(c).reindex(com_i)
 
     # com up
@@ -457,8 +903,8 @@ def define_ramp_limit_constraints(n, sns, c, attr):
         limit_start = assets.eval("ramp_limit_start_up * p_nom").to_xarray()
         limit_up = assets.eval("ramp_limit_up * p_nom").to_xarray()
 
-        status = m[f"{c}-status"].sel(snapshot=active.index)
-        status_prev = m[f"{c}-status"].shift(snapshot=1).sel(snapshot=active.index)
+        status = m[f"{c}-status"].loc[sns, com_i].sel(snapshot=active.index)
+        status_prev = m[f"{c}-status"].loc[sns, com_i].shift(snapshot=1).sel(snapshot=active.index)
 
         lhs = (
             (1, p_actual(com_i)),
@@ -474,7 +920,7 @@ def define_ramp_limit_constraints(n, sns, c, attr):
 
         mask = active.reindex(columns=com_i) & assets.ramp_limit_up.notnull()
         m.add_constraints(
-            lhs, "<=", rhs, name=f"{c}-com-{attr}-ramp_limit_up", mask=mask
+            lhs, "<=", rhs, name=f"{c}-com-non-ext-non-mod-{attr}-ramp_limit_up", mask=mask
         )
 
     # com down
@@ -482,8 +928,8 @@ def define_ramp_limit_constraints(n, sns, c, attr):
         limit_shut = assets.eval("ramp_limit_shut_down * p_nom").to_xarray()
         limit_down = assets.eval("ramp_limit_down * p_nom").to_xarray()
 
-        status = m[f"{c}-status"].sel(snapshot=active.index)
-        status_prev = m[f"{c}-status"].shift(snapshot=1).sel(snapshot=active.index)
+        status = m[f"{c}-status"].loc[sns, com_i].sel(snapshot=active.index)
+        status_prev = m[f"{c}-status"].loc[sns, com_i].shift(snapshot=1).sel(snapshot=active.index)
 
         lhs = (
             (1, p_actual(com_i)),
@@ -500,7 +946,66 @@ def define_ramp_limit_constraints(n, sns, c, attr):
         mask = active.reindex(columns=com_i) & assets.ramp_limit_down.notnull()
 
         m.add_constraints(
-            lhs, ">=", rhs, name=f"{c}-com-{attr}-ramp_limit_down", mask=mask
+            lhs, ">=", rhs, name=f"{c}-com-non-ext-non-mod-{attr}-ramp_limit_down", mask=mask
+        )
+
+    # ----------------------- Committable and Modular Generators (ext and non ext) ----------------------------- #
+    
+    com_i = n.get_committable_i(c)
+    mod_i = n.df(c).query(f"({nominal_attrs[c]}_mod>0)").index
+    com_i = com_i.intersection(mod_i).rename(com_i.name)
+
+    assets = n.df(c).reindex(com_i)
+
+    # com up
+    if not assets.ramp_limit_up.isnull().all():
+        limit_start = assets.eval("ramp_limit_start_up * p_nom_mod").to_xarray()
+        limit_up = assets.eval("ramp_limit_up * p_nom_mod").to_xarray()
+
+        status = m[f"{c}-status"].loc[sns, com_i].sel(snapshot=active.index)
+        status_prev = m[f"{c}-status"].loc[sns, com_i].shift(snapshot=1).sel(snapshot=active.index)
+
+        lhs = (
+            (1, p_actual(com_i)),
+            (-1, p_previous(com_i)),
+            (limit_start - limit_up, status_prev),
+            (-limit_start, status),
+        )
+
+        rhs = rhs_start.reindex(columns=com_i)
+        if is_rolling_horizon:
+            status_start = n.pnl(c)["status"][com_i].iloc[start_i]
+            rhs.loc[sns[0]] += (limit_up - limit_start) * status_start
+
+        mask = active.reindex(columns=com_i) & assets.ramp_limit_up.notnull()
+        m.add_constraints(
+            lhs, "<=", rhs, name=f"{c}-com-mod-{attr}-ramp_limit_up", mask=mask
+        )
+
+    # com down
+    if not assets.ramp_limit_down.isnull().all():
+        limit_shut = assets.eval("ramp_limit_shut_down * p_nom_mod").to_xarray()
+        limit_down = assets.eval("ramp_limit_down * p_nom_mod").to_xarray()
+
+        status = m[f"{c}-status"].loc[sns, com_i].sel(snapshot=active.index)
+        status_prev = m[f"{c}-status"].loc[sns, com_i].shift(snapshot=1).sel(snapshot=active.index)
+
+        lhs = (
+            (1, p_actual(com_i)),
+            (-1, p_previous(com_i)),
+            (limit_down - limit_shut, status),
+            (limit_shut, status_prev),
+        )
+
+        rhs = rhs_start.reindex(columns=com_i)
+        if is_rolling_horizon:
+            status_start = n.pnl(c)["status"][com_i].iloc[start_i]
+            rhs.loc[sns[0]] += -limit_shut * status_start
+
+        mask = active.reindex(columns=com_i) & assets.ramp_limit_down.notnull()
+
+        m.add_constraints(
+            lhs, ">=", rhs, name=f"{c}-com-mod-{attr}-ramp_limit_down", mask=mask
         )