+++ pypsa/optimization/variables.py
@@ -31,7 +31,7 @@ def define_operational_variables(n, sns, c, attr):
     coords = [sns, n.df(c).index.rename(c)]
     n.model.add_variables(coords=coords, name=f"{c}-{attr}", mask=active)
 
-
+#TO BE REMOVED
 def define_status_variables(n, sns, c):
     com_i = n.get_committable_i(c)
 
@@ -46,7 +46,43 @@ def define_status_variables(n, sns, c):
         coords=coords, name=f"{c}-status", mask=active, binary=is_binary, **kwargs
     )
 
+def define_integer_committability_variables(n, sns, c):
+    """
+    Initializes integer variables  related to committability for all committable components. Initialized variables are:
+    - status
+    - start-up
+    - shut down
 
+    Parameters
+    ----------
+    n : pypsa.Network
+    sns : pd.Index
+        Snapshots of the constraint.
+    c : str
+        network component of which the committability variables should be defined
+    """
+    com_i = n.get_committable_i(c)
+
+    if com_i.empty:
+        return
+
+    active = get_activity_mask(n, c, sns, com_i) if n._multi_invest else None
+    coords = (sns, com_i)
+    is_integer = not n._linearized_uc
+    
+    n.model.add_variables(
+        lower=0, coords=coords, name=f"{c}-status",  mask=active, integer=is_integer
+        )
+    
+    n.model.add_variables(
+        lower=0, coords=coords, name=f"{c}-start_up",  mask=active, integer=is_integer
+        )
+    
+    n.model.add_variables(
+        lower=0, coords=coords, name=f"{c}-shut_down",  mask=active, integer=is_integer
+        )
+
+#TO BE REMOVED
 def define_start_up_variables(n, sns, c):
     com_i = n.get_committable_i(c)
 
@@ -61,7 +97,7 @@ def define_start_up_variables(n, sns, c):
         coords=coords, name=f"{c}-start_up", mask=active, binary=is_binary, **kwargs
     )
 
-
+#TO BE REMOVED
 def define_shut_down_variables(n, sns, c):
     com_i = n.get_committable_i(c)